% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{appointments}
\alias{appointments}
\title{patient appointment list}
\usage{
appointments(dMeasureAppointments_obj)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}
}
\value{
dataframe of apppointments
$Patient, $AppointmentDate, $AppointmentTime,
$Provider, $Status
}
\description{
derived from dM$appointments_filtered_time
}
