% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{appointments_telephone}
\alias{appointments_telephone}
\title{patient appointment list with telephone numbers}
\usage{
appointments_telephone(dMeasureAppointments_obj)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}
}
\value{
dataframe of apppointments
$Patient, $AppointmentDate, $AppointmentTime,
$Provider, $Status, $MobilePhone, $HomePhone, $WorkPhone
}
\description{
derived from dM$appointments_filtered_time
}
