% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_userInterface.R
\name{datatableServer}
\alias{datatableServer}
\title{Appointments module - server}
\usage{
datatableServer(id, dMAppointments)
}
\arguments{
\item{id}{id}

\item{dMAppointments}{dMeasureAppointments R6 object}
}
\value{
none
}
\description{
Appointments module - server
}
