% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_userInterface_configuration.R
\name{dMeasureConfigurationTabPanel}
\alias{dMeasureConfigurationTabPanel}
\title{Appointments module - configuration panel server}
\usage{
dMeasureConfigurationTabPanel(id, dMAppointments)
}
\arguments{
\item{id}{as required by Shiny modules}

\item{dMAppointments}{dMeasureAppointments R6 object}
}
\value{
none
}
\description{
Appointments module - configuration panel server
}
\details{
This (module) server is called by `DailyMeasure` with the
  same module `id` as `dMeasureConfigurationTabPanelItem` will
  call `dMeasureConfigurationTabPanelUI`

  The module `id` is also returned by the integration function
  `dMeasureIntegration` when called with the parameter `configID`
}
