% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_userInterface_configuration.R
\name{dMeasureConfigurationTabPanelItem}
\alias{dMeasureConfigurationTabPanelItem}
\title{Configuration interface elements of dMeasureAppointments}
\usage{
dMeasureConfigurationTabPanelItem()
}
\value{
tabPanel
}
\description{
requires R6 methods from fct_integration.R
}
\details{
This function is called by `DailyMeasure`
}
