% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_userInterface_configuration.R
\name{dMeasureConfigurationTabPanelUI}
\alias{dMeasureConfigurationTabPanelUI}
\title{Appointments module - configuration panel UI}
\usage{
dMeasureConfigurationTabPanelUI(id)
}
\arguments{
\item{id}{module ID}
}
\value{
shiny user interface element
}
\description{
Appointments module - configuration panel UI
}
\details{
This function is called by `dMeasureConfigurationTabPanelItem`,
  with the same module `id` as `DailyMeasure` will
  call `dMeasureConfigurationTabPanel`.

  The module `id` is also returned by the integration function
  `dMeasureIntegration` when called with parameter `configID`
}
