% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_integration.R
\name{initialize_join_config}
\alias{initialize_join_config}
\title{dMeasureAppointment module - initialize Join configuration}
\usage{
initialize_join_config()
}
\value{
list
}
\description{
dMeasureAppointment module - initialize Join configuration
}
