% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_integration.R
\name{initialize_sms_config}
\alias{initialize_sms_config}
\title{dMeasureAppointment module - initialize SMS text configuration}
\usage{
initialize_sms_config()
}
\value{
list
}
\description{
dMeasureAppointment module - initialize SMS text configuration
}
