% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_integration.R
\name{read_configuration_db}
\alias{read_configuration_db}
\title{read the configuration database}
\usage{
read_configuration_db(dMeasureAppointment_obj)
}
\arguments{
\item{dMeasureAppointment_obj}{R6 object}
}
\value{
Join configuration
}
\description{
read the configuration database
}
\details{
reads join configuration.
  reads SMS configuration.
}
