% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_configuration.R
\name{remove_join_config}
\alias{remove_join_config}
\title{remove join definition from configuration database}
\usage{
remove_join_config(dMeasureAppointments_obj, api)
}
\arguments{
\item{api}{name of API (at this stage, can only be 'Join')}

\item{Appointments_obj}{R6 object}
}
\value{
TRUE if removed successfully
}
\description{
remove join definition from configuration database
}
