% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_configuration.R
\name{remove_sms_config}
\alias{remove_sms_config}
\title{remove SMS text definition from configuration database}
\usage{
remove_sms_config(dMeasureAppointments_obj, name)
}
\arguments{
\item{name}{name of SMS text}

\item{Appointments_obj}{R6 object}
}
\value{
TRUE if removed successfully
}
\description{
remove SMS text definition from configuration database
}
