% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_integration.R
\name{sidebarmenuPriority}
\alias{sidebarmenuPriority}
\title{sidebarmenuPriority}
\usage{
sidebarmenuPriority()
}
\value{
priority.
}
\description{
priority order to appear in sidebarmenu
  integration with DailyMeasure

50 is medium priority. larger numbers are higher priority
}
