% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sms.R
\name{sms_join}
\alias{sms_join}
\title{sms_join}
\usage{
sms_join(smstext, smsnumber, apikey = NULL, deviceId = NULL)
}
\arguments{
\item{smstext}{the text to send}

\item{smsnumber}{the sms number (in characters)}

\item{apikey}{apikey to Join}

\item{deviceId}{deviceId of the mobile device}
}
\value{
interpreted response object
}
\description{
Send SMS with Join
}
\details{
Join (by joaoapps) allows sending SMSs with an Android phone connected to a Google account
https://joaoapps.com/join/

Instructions to find \code{apkey} and \code{deviceId} can be found at http://www.juanmtech.com/join-by-joaoapps-beginners-guide/

In summary (after device has Join installed and connected to a Google account)
\itemize{
\item Go to Join's web interface at https://joinjoaomgcd.appspot.com/, and login with your Google account
\item Choose the device which you will use to send SMSs
\item Choose 'Create a Join API for your device' from the right-hand-side list of options
\item The 'Device Id' and 'API Key' will be visible at the bottom of the web page
}

The returned response objected is interpreted by \code{httr::content}

\verb{$success} will be \code{TRUE} or \code{FALSE}

If there is an error (\verb{$success == FALSE}), then \verb{$errorMessage} will contain an error message such as
\itemize{
\item "User not Authenticated" (incorrect \code{apikey})
\item "No device to send message to" (incorrect \code{deviceId})
}
}
