% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_configuration.R
\name{update_join_config}
\alias{update_join_config}
\title{update join_config item in configuration database}
\usage{
update_join_config(dMeasureAppointments_obj, id, api, apikey, deviceId)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}

\item{id}{row ID}

\item{api}{name of API}

\item{deviceId}{}
}
\value{
ID of patient list.
 stop errors generated for several different reasons of failure
}
\description{
update join_config item in configuration database
}
