% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_configuration.R
\name{update_sms_config}
\alias{update_sms_config}
\title{update sms_config item in configuration database}
\usage{
update_sms_config(dMeasureAppointments_obj, id, name, smstext)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}

\item{id}{row ID}

\item{name}{name of SMS text}

\item{smstext}{}
}
\value{
ID of patient list.
 stop errors generated for several different reasons of failure
}
\description{
update sms_config item in configuration database
}
