% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_configuration.R
\name{write_join_config}
\alias{write_join_config}
\title{Configuration elements of dMeasureAppointments}
\usage{
write_join_config(dMeasureAppointments_obj, api, apikey, deviceId)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}

\item{api}{name of patient list}

\item{deviceId}{}
}
\value{
ID of created item
}
\description{
write to configuration database methods
}
\details{
requires R6 methods from fct_integration.R
}
