% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_configuration.R
\name{write_sms_config}
\alias{write_sms_config}
\title{insert sms_config item to configuration database}
\usage{
write_sms_config(dMeasureAppointments_obj, name, smstext)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}

\item{name}{name of SMS text}

\item{smstext}{}
}
\value{
ID of created item
}
\description{
insert sms_config item to configuration database
}
