% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{appointments}
\alias{appointments}
\title{patient appointment list}
\usage{
appointments(
  dMeasureAppointments_obj,
  screentag = FALSE,
  screentag_print = TRUE
)
}
\arguments{
\item{dMeasureAppointments_obj}{R6 object}

\item{screentag}{if neither \code{screentag} or \code{screentag_print} is
defined then values are derived from \code{self$printcopy_view}}

\item{screentag_print}{}
}
\value{
dataframe of apppointments
$Patient, $AppointmentDate, $AppointmentTime,
$Provider, $Status
}
\description{
derived from dM$appointments_filtered_time
}
