% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_integration.R
\name{dMeasureAppointments}
\alias{dMeasureAppointments}
\title{dMeasureAppointments}
\description{
dMeasureAppointments

list appointments by clinician providers
}
\details{
dMeasureAppointments class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dMeasureAppointments$new()}}
\item \href{#method-read_configuration_db}{\code{dMeasureAppointments$read_configuration_db()}}
\item \href{#method-appointments}{\code{dMeasureAppointments$appointments()}}
\item \href{#method-appointments_telephone}{\code{dMeasureAppointments$appointments_telephone()}}
\item \href{#method-clone}{\code{dMeasureAppointments$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureAppointments$new(dMeasure_obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_configuration_db"></a>}}
\if{latex}{\out{\hypertarget{method-read_configuration_db}{}}}
\subsection{Method \code{read_configuration_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureAppointments$read_configuration_db()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-appointments"></a>}}
\if{latex}{\out{\hypertarget{method-appointments}{}}}
\subsection{Method \code{appointments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureAppointments$appointments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-appointments_telephone"></a>}}
\if{latex}{\out{\hypertarget{method-appointments_telephone}{}}}
\subsection{Method \code{appointments_telephone()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureAppointments$appointments_telephone()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dMeasureAppointments$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
